Split ToolTipFunctions into two files
